aiModel/requirements.txt
fastapi>=0.115.2
uvicorn>=0.24.0
python-multipart>=0.0.18
pillow>=10.1.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.35.0
numpy<2.0.0
pika>=1.3.2
rembg>=2.0.55
diffusers>=0.21.4
accelerate>=0.24.1
huggingface-hub>=0.19.3

# Google AI Studio (Gemini 2.5 Flash Image - Nano Banana)
google-generativeai>=0.8.0
python-dotenv>=1.0.0
